refactor(supplier-profile): extract ProductBasicsStep into its own file
